pyproject.toml
[project]
name = "ursa-ai"
dynamic = ["version"]
description = "Agents for science at LANL"
readme = "README.md"
authors = [
    { name = "Mike Grosskopf", email = "mikegros@lanl.gov" },
    { name = "Nathan Debardeleben", email = "ndebard@lanl.gov" },
    { name = "Rahul Somasundaram", email = "rsomasundaram@lanl.gov" },
    { name = "Isaac Michaud", email = "imichaud@lanl.gov" },
    { name = "Avanish Mishra", email = "avanish@lanl.gov" },
    { name = "Arthur Lui", email = "alui@lanl.gov" },
    { name = "Russell Bent", email = "rbent@lanl.gov" },
    { name = "Earl Lawrence", email = "earl@lanl.gov" }
]
license = "BSD-3-Clause"
license-files = [ "LICENSE" ]
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0,<3.0",
    "beautifulsoup4>=4.13.4,<5.0",
    "coolname>=2.2.0,<3.0",
    "langchain>=0.3.27,<0.4",
    "langchain-community>=0.3.27,<0.4",
    "langchain-litellm>=0.2.2,<0.3",
    "langchain-openai>=0.3.30,<0.4",
    "langgraph>=0.6.6,<0.7",
    "pandas>=2.3.1,<3.0",
    "pillow>=11.3.0,<12.0",
    "pymupdf>=1.26.0,<2.0",
    "pypdf>=5.9.0,<6.0",
    "rich>=13.9.4,<14.0",
    "langchain-chroma>=0.2.5,<0.3",
    "chromadb>=1.0.20,<1.1",
    "mp-api>=0.45.8,<0.46",
    "langchain-google-genai>=2.1.9,<3.0",
    "langchain-anthropic>=0.3.19,<0.4",
    "langgraph-checkpoint-sqlite>=2.0.10,<3.0",
    "duckduckgo-search>=8.1,<9.0",
    "langchain-ollama>=0.3.6,<0.4",
]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.urls]
Homepage = "https://github.com/lanl/ursa"
Documentation = "https://github.com/lanl/ursa/tree/main/docs"
Repository = "https://github.com/lanl/ursa"
Issues = "https://github.com/lanl/ursa/issues"

[build-system]
requires = ["setuptools>=74.1", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

[tool.setuptools-git-versioning]
enabled = true

[tool.ruff]
line-length = 80

[tool.ruff.lint]
ignore = ["D100"]
extend-select = ["I", "W505"] # "D"
extend-unsafe-fixes = ["F401"]
pydocstyle.convention = "numpy"
pycodestyle.max-doc-length = 80

# Ignore test file documentation linting.
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["D"]

[dependency-groups]
dev = [
    "langgraph-checkpoint-sqlite>=2.0.10",
    "notebook>=7.3.3",
    "pre-commit>=4.3.0",
    "scikit-optimize>=0.10.2",
]
